+++ tests/unit/resources/settings/test_appliance_device_snmp_v3_trap_destinations.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2021] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -21,7 +21,7 @@ import mock
 
 from hpeOneView.connection import connection
 from hpeOneView.resources.settings.appliance_device_snmp_v3_trap_destinations import ApplianceDeviceSNMPv3TrapDestinations
-from hpeOneView.resources.resource import ResourceClient
+from hpeOneView.resources.resource import Resource, ResourceHelper
 
 
 class ApplianceDeviceSNMPv3TrapDestinationsTest(unittest.TestCase):
@@ -29,13 +29,50 @@ class ApplianceDeviceSNMPv3TrapDestinationsTest(unittest.TestCase):
         self.host = '127.0.0.1'
         self.connection = connection(self.host, 800)
         self._snmp_v3_trap_dest = ApplianceDeviceSNMPv3TrapDestinations(self.connection)
+        self.uri = "/rest/appliance/snmpv3-trap-forwarding/destinations"
+        self._snmp_v3_trap_dest.data = {"uri": self.uri}
 
-    @mock.patch.object(ResourceClient, 'get_all')
-    def test_get_all_called_once(self, mock_get):
+    @mock.patch.object(ResourceHelper, 'get_all')
+    def test_get_all_called_once(self, mock_get_all):
         self._snmp_v3_trap_dest.get_all()
-        mock_get.assert_called_once_with(0, -1, filter='', sort='')
+        mock_get_all.assert_called_once_with(0, -1, filter='', sort='', query='')
 
-    @mock.patch.object(ResourceClient, 'create')
+    @mock.patch.object(ResourceHelper, 'get_all')
+    def test_get_all_called_once_with_default(self, mock_get_all):
+        self._snmp_v3_trap_dest.get_all()
+        mock_get_all.assert_called_once_with()
+
+    @mock.patch.object(Resource, 'get_by_uri')
+    def test_get_by_uri_called_once(self, mock_get_by_uri):
+        uri = "/rest/appliance/snmpv3-trap-forwarding/destinations/1"
+        self._snmp_v3_trap_dest.get_by_uri(uri)
+        mock_get_by_uri.assert_called_once_with(uri)
+
+    # @mock.patch.object(ResourceHelper, 'create')
+    # @mock.patch.object(Resource, 'create')
+    # def test_create_called_once(self, mock_create, mock_create_validation):
+    #     create_uri = "{}/1".format(self.uri)
+    #     validation_uri = "{}/validation".format(self.uri)
+    #     resource = {
+    #         'type': 'Destination',
+    #         'destinationAddress': '1.1.1.1',
+    #         'port': 162,
+    #         'userId': '6b9c6f7b-7a24-4514-b9c9-0c31e086c170'
+    #     }
+    #
+    #     resource_validation = {
+    #         'destinationAddress': '1.1.1.1',
+    #     }
+    #
+    #     resource_rest_call = resource.copy()
+    #     mock_create.return_value = {}
+    #     mock_create_validation.return_value = {}
+    #
+    #     self._snmp_v3_trap_dest.create(resource)
+    #     mock_create_validation.assert_called_once_with(resource_validation, uri=validation_uri, timeout=-1)
+    #     mock_create.assert_called_once_with(resource_rest_call, uri=create_uri, timeout=-1)
+
+    @mock.patch.object(Resource, 'create')
     def test_create_called_once(self, mock_create):
         resource = {
             'type': 'Destination',
@@ -46,35 +83,41 @@ class ApplianceDeviceSNMPv3TrapDestinationsTest(unittest.TestCase):
         self._snmp_v3_trap_dest.create(resource)
         mock_create.assert_called_once_with(resource, timeout=-1)
 
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_called_once(self, mock_get):
-        self._snmp_v3_trap_dest.get('86731e55-6837-44cf-a5c5-f0392920da7e')
-
-        mock_get.assert_called_once_with('86731e55-6837-44cf-a5c5-f0392920da7e')
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_with_uri_called_once(self, mock_get):
-        uri = '/rest/appliance/snmpv3-trap-forwarding/destinations/86731e55-6837-44cf-a5c5-f0392920da7e'
-        self._snmp_v3_trap_dest.get(uri)
-
-        mock_get.assert_called_once_with(uri)
-
-    @mock.patch.object(ResourceClient, 'get_by')
-    def test_get_by_uri_called_once(self, mock_create):
-        uri = '/rest/appliance/snmpv3-trap-forwarding/destinations/86731e55-6837-44cf-a5c5-f0392920da7e'
-        self._snmp_v3_trap_dest.get_by('uri', uri)
-        mock_create.assert_called_once_with('uri', uri)
-
-    @mock.patch.object(ResourceClient, 'update')
-    def test_update_called_once(self, mock_create):
+    @mock.patch.object(ResourceHelper, 'create')
+    def test_create_validation_called_once(self, mock_create_validation):
         resource = {
-            'destinationAddress': '1.1.9.9',
+            'destinationAddress': '1.1.1.1',
+            'existingDestinations': ['1.2.3.4']
+        }
+
+        validation_uri = "{}/validation".format(self.uri)
+        resource_rest_call = resource.copy()
+        mock_create_validation.return_value = {}
+
+        self._snmp_v3_trap_dest.create_validation(resource['destinationAddress'], resource['existingDestinations'])
+        mock_create_validation.assert_called_once_with(resource_rest_call, uri=validation_uri, timeout=-1)
+
+    @mock.patch.object(Resource, 'ensure_resource_data')
+    @mock.patch.object(ResourceHelper, 'update')
+    def test_update_called_once_with_default(self, mock_update, mock_ensure_client):
+        resource = {
+            "destinationAddress": "1.1.1.2",
+            "uri": self.uri
         }
         self._snmp_v3_trap_dest.update(resource)
-        mock_create.assert_called_once_with(resource, timeout=-1)
+        mock_update.assert_called_once_with(resource, self.uri, False, -1, None)
 
-    @mock.patch.object(ResourceClient, 'delete')
-    def test_delete_called_once(self, mock_create):
-        id_or_uri = '/rest/appliance/snmpv3-trap-forwarding/destinations/86731e55-6837-44cf-a5c5-f0392920da7e'
-        self._snmp_v3_trap_dest.delete(id_or_uri)
-        mock_create.assert_called_once_with(id_or_uri, timeout=-1)
+    @mock.patch.object(Resource, 'ensure_resource_data')
+    @mock.patch.object(ResourceHelper, 'update')
+    def test_update_called_once(self, mock_update, mock_ensure_client):
+        resource = {
+            "uri": self.uri,
+            "destinationAddress": "1.1.1.2",
+        }
+        self._snmp_v3_trap_dest.update(resource, 70)
+        mock_update.assert_called_once_with(resource, self.uri, False, 70, None)
+
+    @mock.patch.object(Resource, 'delete')
+    def test_delete_called_once(self, mock_delete):
+        self._snmp_v3_trap_dest.delete()
+        mock_delete.assert_called_once_with()